+++ __tests__/junit/gotestsum.test.ts
@@ -1,6 +1,6 @@
 import * as fs from 'fs'
 
-import { JunitReport, TestResult, TestCase } from '../../src/junit/report'
+import { GotestsumReport, TestResult, TestCase } from '../../src/junit/gotestsum'
 
 describe('reporter', () => {
   it('parses the junit report', async () => {
@@ -28,7 +28,7 @@ describe('reporter', () => {
       </testsuites>
       `)
 
-    const report = await JunitReport.fromXml('path/to/junit.xml')
+    const report = await GotestsumReport.fromXml('path/to/junit.xml')
     expect(report.directory).toBe('path/to')
     expect(report.result).toBe(TestResult.Failed)
     expect(report.tests).toBe(4)
@@ -53,7 +53,7 @@ describe('reporter', () => {
   })
 
   it('constructs a unknown report', async () => {
-    const report = JunitReport.unknown('path/to/junit.xml')
+    const report = GotestsumReport.unknown('path/to/junit.xml')
     expect(report.directory).toBe('path/to')
     expect(report.result).toBe(TestResult.Unknown)
     expect(report.tests).toBe(0)
