+++ src/monorepo.ts
@@ -1,6 +1,6 @@
 import glob from 'fast-glob'
 
-import { Reportable, JunitReport, TestResult } from './junit/report'
+import { Reportable, GotestsumReport, TestResult } from './junit/gotestsum'
 import path from 'path'
 
 export type MarkdownContext = RepositoryContext &
@@ -37,7 +37,7 @@ export class Monorepo {
   ): Promise<Monorepo> {
     const files = directories.map(directory => path.join(directory, filename))
     const reporters = await Promise.all(
-      files.map(async file => await JunitReport.fromXml(file))
+      files.map(async file => await GotestsumReport.fromXml(file))
     )
     return new Monorepo(reporters)
   }
@@ -45,7 +45,7 @@ export class Monorepo {
   static async fromFilename(filename: string): Promise<Monorepo> {
     const files = await glob(`**/${filename}`, { dot: true })
     const reporters = await Promise.all(
-      files.map(async file => await JunitReport.fromXml(file))
+      files.map(async file => await GotestsumReport.fromXml(file))
     )
     return new Monorepo(reporters)
   }
